feature-policy/experimental-features/lazyload/lazyload-disabled-tentative.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/feature-policy/experimental-features/resources/common.js"></script>
<style>
html, body {
  height: 100%;
  width: 100%;
}

iframe {
  width: 400px;
  height: 400px;
  margin: 10px;
}

.spacer {
  width: 100%;
  height: 300%;
}
</style>
<div class="spacer"></div>
<iframe id="frame-off" lazyload="off"></iframe>
<iframe id="frame-auto" lazyload="auto"></iframe>
<iframe id="frame-on" lazyload="on"></iframe>

<script>
  let cross_origin_url =
      "http://{{hosts[alt][www]}}:{{ports[http][0]}}/" +
      "feature-policy/experimental-features/resources/lazyload-contents.html";

  let frameOff = document.getElementById("frame-off");
  let frameAuto = document.getElementById("frame-auto");
  let frameOn = document.getElementById("frame-on");
  let load_timeout = 100; // ms
  let expected_timeout_msg = false;

  window.scrollTo(0, 0);

  // Verify the page is properly initialized.
  test(() => {
    [frameOff, frameAuto, frameOn].forEach((f) => {
      assert_true(!!f);
      let top = f.getBoundingClientRect().top;
      assert_greater_than(top, window.innerHeight * 2,
          `Unexpected position for frame id=${f.getAttribute('id')}.`);
    });
  }, "Make sure the <iframe> is not visible.");

  // Verify that lazyload='on|off|auto' work as expected ('off' and 'auto'
  // should lead to the same result).
  promise_test(async(t) => {
    frameOn.src = `${cross_origin_url}?id=frame-on`;
    await waitForMessageOrTimeout(t, "frame-on", load_timeout).then((msg) => {
      assert_equals(msg,
        expected_timeout_msg,
        "With lazyload='on', the frame should not load.");
    });
    frameAuto.src = `${cross_origin_url}?id=frame-auto`;
    await waitForMessageOrTimeout(t, "frame-auto", load_timeout).then(
      (msg_auto) => {
        frameOff.src = `${cross_origin_url}?id=frame-off`;
        return waitForMessageOrTimeout(t, "frame-off", load_timeout).then(
          (msg_off) => {
            assert_equals(
              msg_auto,
              msg_off,
              "'off' and 'auto' should behave the same.");
        });
    });
  }, "Sanity-check: 'lazyload attribute' works as expected when the " +
     "'lazyload feature policy' is disabled.");
</script>
